Rebalance PostOperatorCSV ctor and InitializeCSVDataCollection
- Move fem_op_t so PostOperatorCSV ctor can use it
